requirements.txt
beautifulsoup4==4.12.2
emoji==2.2.0
Flask==2.2.2
geopy==2.3.0
icalendar==5.0.4
matplotlib==3.7.1
mysql_connector_repackaged==0.3.1
nltk==3.8.1
numpy==1.24.1
pandas==1.5.3
pytz==2022.7.1
regex==2023.3.23
requests==2.28.2
scikit_learn==1.2.2
seaborn==0.12.2
SQLAlchemy==2.0.1
yelpapi==2.5.0
